refac: add js modules
