+++ museum-dom/index.html
@@ -1,6 +1,8 @@
 <!-- TODO: add arrows hover -->
 <!-- TODO: random gallery pics positions -->
 
+<!-- TODO: mouse drag moves slide and linear transition -->
+
 <!DOCTYPE html>
 <html lang="en">
   <head>
@@ -1221,6 +1223,6 @@
       <!-- /.footer__bottom -->
     </footer>
 
-    <script type="module" src="js/script.js"></script>
+    <script type="module" src="js/index.js"></script>
   </body>
 </html>
