+++ museum-dom/js/video.js
@@ -1,0 +1,7 @@
+//progress-bar in video
+function progressAnimation() {
+  const value = this.value;
+  this.style.background = `linear-gradient(to right, #710707 0%, #710707 ${value}%, #c4c4c4 ${value}%, #c4c4c4 100%)`;
+}
+
+export { progressAnimation };
